moved the filter components into one huge component + fixed responsiveness
