+++ frontend-v2/src/components/SearchBar/SearchBar.tsx
@@ -2,7 +2,7 @@
 
 import { MagnifyingGlassIcon } from "@heroicons/react/24/solid";
 import { FormEvent, useState } from "react";
-import FilterSelection from "../FilterSelection/FilterSelection";
+import FilterModal from "../FilterModal/FilterModal";
 
 export default function SearchBar() {
   const handleOnSubmit = (event: FormEvent) => {
@@ -13,9 +13,9 @@ export default function SearchBar() {
 
   return (
     <div className="flex w-5/6 items-center bg-white rounded border-2 border-unilectives-search">
-      <form className="flex w-full" onSubmit={handleOnSubmit}>
+      <form className="flex px-2 items-center w-full" onSubmit={handleOnSubmit}>
         <button type="submit">
-          <MagnifyingGlassIcon className="w-6 h-6 text-unilectives-search mx-2" />
+          <MagnifyingGlassIcon className="w-6 h-6 text-unilectives-search" />
         </button>
         <input
           type="text"
@@ -24,8 +24,7 @@ export default function SearchBar() {
           placeholder="Search for a course e.g. COMP1511"
         />
       </form>
-      <FilterSelection />
+      <FilterModal />
     </div>
   );
 }
-
